+++ NEWS.md
@@ -23,6 +23,7 @@
   * `author` looks at `options("openxlsx2.creator")` in `wb_comment()` compared to only `sys.getenv("user")` in `create_comment()`
   * `visible` defaults to `FALSE` in `wb_comment()` to account for modern spreadsheet software behaviour.
   In `create_comment()`, it is `TRUE`.
+  * `width` and `height` must now be of length 1. (In `create_comment()`, the first element is taken, other are ignored.)
 
 ## Internal changes